HTBHF-2499 update session ID checks to operate on the first page of the application, not the guidance page
